+++ src/Hellion.Login/Client/LoginClient.cs
@@ -70,6 +70,8 @@ namespace Hellion.Login.Client
             {
                 case LoginHeaders.Incoming.LoginRequest: this.OnLoginRequest(pak); break;
 
+
+
                 default: FFPacket.UnknowPacket<LoginHeaders.Incoming>((uint)packetHeaderNumber, 2); break;
             }
         }