+++ CHANGELOG.md
@@ -8,6 +8,7 @@
         - Map loading (.dyo / .rgn / .wld)
     - Player login
     - Player visibility with other players
+    - Player moves
 
 ### Update #1 (09/12/2016)